renamed classifier_api to mentor-api
This reflects the emerging reality that what has been called
classifier-api does more than just run the mentor classifiers.

For now, it is the data api for all things mentor,
including things like topic data files.

Down the road, we should remove the running of the classifier
itself from mentor-api and have it run in a worker service
and communicate with mentor-api via, say, celery (redis mq)
