+++ bin/kubernetes_deploy.sh
@@ -5,10 +5,10 @@ KUBERNETES_DIR=${1:-./build/kubernetes}
 echo "Creating the secrets store..."
 kubectl apply -f ${KUBERNETES_DIR}/secrets.yml
 
-echo "Creating the classifier-api deployment and service..."
+echo "Creating the mentor-api deployment and service..."
 
-kubectl apply -f ${KUBERNETES_DIR}/classifier-api-deployment.yml
-kubectl apply -f ${KUBERNETES_DIR}/classifier-api-service.yml
+kubectl apply -f ${KUBERNETES_DIR}/mentor-api-deployment.yml
+kubectl apply -f ${KUBERNETES_DIR}/mentor-api-service.yml
 
 echo "Creating the web_app deployment and service..."
 kubectl apply -f ${KUBERNETES_DIR}/web-app-deployment.yml
