--- services/classifier_api/src/manage.py
@@ -1,3 +1,0 @@
-from mentorpal_classifier_api import create_app
-
-app = create_app()
