--- services/classifier_api/tests/flask_stop.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-CONTAINER_NAME=mentorpal_classifier_api_testing
-
-docker stop ${CONTAINER_NAME}
