+++ services/mentor-api/src/manage.py
@@ -1,0 +1,3 @@
+from mentor_api import create_app
+
+app = create_app()
