+++ services/mentor-api/tests/flask_start.sh
@@ -5,8 +5,8 @@ CLASSIFIER_API=$(dirname ${TESTS})
 SERVICES=$(dirname ${CLASSIFIER_API})
 ROOT=$(dirname ${SERVICES})
 
-IMAGE_NAME=uscictdocker/mentorpal-classifier-api:latest
-CONTAINER_NAME=mentorpal_classifier_api_testing
+IMAGE_NAME=uscictdocker/mentorpal-mentor-api:latest
+CONTAINER_NAME=mentorpal-mentor-api-testing
 
 # The flask docker image/api we're running
 # should be the same one used in production envs,
@@ -20,7 +20,7 @@ docker run \
 		--rm \
 		--name ${CONTAINER_NAME} \
 		-p 5000:5000 \
-		-v ${CLASSIFIER_API}/src/mentorpal_classifier_api:/app/mentorpal_classifier_api \
+		-v ${CLASSIFIER_API}/src/mentorpal_mentor_api:/app/mentorpal_mentor_api \
 		-v ${ROOT}/src/mentorpal:/app/mentorpal \
 		-v ${DOCKER_MOUNT_SRC}:${DOCKER_MOUNT_TGT} \
 		-e MENTORPAL_CLASSIFIER_API_SETTINGS=${FLASK_CONFIG_TGT} \
