+++ services/mentor-api/tests/flask_stop.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+CONTAINER_NAME=mentorpal-mentor-api-testing
+
+docker stop ${CONTAINER_NAME}
