+++ services/proxy/conf.d/default.conf
@@ -5,7 +5,7 @@ server {
 
 
     location ~* /mentor-api {
-        proxy_pass http://classifier-api:5000;
+        proxy_pass http://mentor-api:5000;
         proxy_http_version 1.1;
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection 'upgrade';
